Create 'k_sequence.c/h' files + implement 'K_Seq_DungeonCall', 'K_Seq_DungeonExit' and 'K_Seq_DungeonCheck'

+++ include/kwln/kwlnTask.h
@@ -64,7 +64,7 @@ u8 kwlnTaskDestroyWithHierarchy(KwlnTask* task);
 u32 kwlnTaskGetState(KwlnTask* task);
 KwlnTask* kwlnTaskGetTaskByName(const char* name);
 KwlnTask* kwlnTaskGetUpdating();
-u8 kwlnTaskExists(KwlnTask* task);
+u32 kwlnTaskExists(KwlnTask* task);
 u32 kwlnTaskGetTimer(KwlnTask* task);
 void kwlnTaskSetWorkData(KwlnTask* task, void* workData);
 void* kwlnTaskGetWorkData(KwlnTask* task);
